requirements.txt
Django==1.10
django-allauth==0.31.0
django-braces==1.11.0
django-cors-headers==2.0.2
django-crispy-forms==1.6.1
django-extensions==1.8.1
django-filter==1.0.2
django-oauth-toolkit==0.12.0
django-registration==2.2
django-rest-auth==0.9.1
django-reversion==2.0.8
django-star-ratings==0.5.6
djangorestframework==3.6.2
djangorestframework-jwt==1.11.0
drfdocs==0.0.11
gunicorn==19.7.1
Pillow==4.0.0
prompt-toolkit==1.0.14
psycopg2-binary==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
PyJWT==1.5.2
pyparsing==2.1.10
python-dateutil==2.6.1
python-openid==2.2.5
python-whois==0.6.5
pytz==2016.10
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.13.0
requests-oauthlib==0.8.0
scandir==1.5
scikit-learn==0.19.1
scipy==0.19.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
subprocess32==3.2.7
sympy==1.1
terminado==0.6
testpath==0.3.1
tornado==4.5.1
traitlets==4.3.2
vine==1.1.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==2.0.0
